+++ src/taco-cli/resources/en/resources.json
@@ -27,8 +27,6 @@
     "_CommandCreateNotBothTemplateAndCordpva.comment": "Error shown when the user tries to create a project using a template (--template), but specifies the --cordova parameter.",
     "CommandCreateTemplatesUnavailable": "Error: Unable to access or download templates",
     "_CommandCreateTemplatesUnavailable.comment": "Error reported when templates cannot be downloaded",
-    "CommandCreateOnlyLocalCustomWww": "Error: Only local paths for custom www assets are supported.",
-    "_CommandCreateOnlyLocalCustomWww.comment": "Error reported when the specified path to the --copy-from or --link-to parameters is a web URL.",
     "CommandCreateInstallGulp": "<br/>Please install gulp and gulp-typescript by following the installation steps listed at <link>http://go.microsoft.com/fwlink/?LinkId=620942</link>.<br/><br/>Once installed, run 'gulp typescript' to pre-compile the TS files before building your project.<br/><br/>Once your TS files are pre-compiled, use 'taco build [platform]' to build your project.",
     "_CommandCreateInstallGulp.comment": "Message shown at the end of project creation for typescript projects, asking the user to install gulp and gulp-typescript in order to precompile their typescript files.",
     "CommandCreatePathNotEmpty": "Error: Could not create the project because the specified path already exists and is not empty: {0}",
@@ -69,9 +67,6 @@
     "_CommandCreateSuccessProjectCustomWww.comment": "Success message when the create command succeeds if the project was created using custom www assets. {0} is the path to the custom www assets provided by the user and {1} is the location where project is created.",
     "CommandCreateSuccessProjectCLI": "<br/><success>Success!</success> Your project is ready for use at <highlight>{0}</highlight>",
     "_CommandCreateSuccessProjectCLI.comment": "Success message shown if the project was created using the --cordova parameter. {0} is the location of created project",
-    "CommandCreateTacoJsonFileCreationError": "Error: Creation of taco.json file failed",
-    "_CommandCreateTacoJsonFileCreationError.comment": "This message is displayed when the taco.json file creation fails.",
-    "CommandCreateTacoJsonFileWriteError": "Failed to write to {0}",
 
     "CommandTemplatesDescription": "List available TACO templates for project creation",
     "_CommandTemplatesDescription.comment": "Description of the 'taco templates' command when the user types 'taco help templates'.",
@@ -398,8 +393,6 @@
     "CommandInstallReqsPlatformDescription": "A platform for which to install third-party dependencies",
     "_CommandInstallReqsPlatformDescription.comment": "Description for the [PLATFORM] option(s) of 'taco install-reqs'",
 
-    "CommandInstallCordovaTooOld": "Error: The current project uses Cordova version {0}. 'taco install-reqs' supports projects using Cordova version {1} or higher.<br/><br/>If your project was created using a Kit, consider moving to a kit that has a more recent version of Cordova. If your project was created using Cordova directly, consider updating the installation of Cordova. For more information, please refer to http://go.microsoft.com/fwlink/?LinkId=618544.",
-    "_CommandInstallCordovaTooOld.comment": "Error shown when a user uses 'taco install-reqs' with a Cordova version older than the minimum version. {0} is the Cordova version used by the current project, {1} is the minimum required Cordova version.",
     "CommandInstallNoPlatformsAdded": "Error: No platforms added to your project. To install the dependencies for a given platform, please add it to your project first.",
     "_CommandInstallNoPlatformsAdded.comment": "Error shown when a user uses 'taco install-reqs' but there are no platforms added to the current project.",
     "CommandInstallNoPlatformsFolder": "Error: No 'platforms' folder found. Please make sure the current directory is the root of a Cordova project, and at least one platform has been added.",
@@ -428,9 +421,6 @@
 
     "RemoteBuildUnsupportedPlatform": "Error: Platform '{0}' not supported by current version of Cordova.",
     "_RemoteBuildUnsupportedPlatformcomment": "Error message printed when trying to 'taco remote add <platform>' and the platform is not supported. {0} is the platform",
-
-    "InvalidCordovaWithNode5": "Incompatible configuration: Unable to use Cordova < 5.4.0 with Node >= 5.0.0. Either upgrade to a newer version of Cordova, or revert to an older version of node. See http://go.microsoft.com/fwlink/?LinkID=618471 for more details.",
-    "_InvalidCordovaWithNode5.comment": "Error message printed when trying to use Cordova < 5.4.0 with Node >= 5.0.0.",
 
     "UnsupportedHostPlatform": "Error: Unsupported platform {0}",
     "_UnsupportedHostPlatform.comment": "Message printed when running on an operating system that we don't support. {0} is the platform",
@@ -534,17 +524,6 @@
     "CleaningRemoteResources": "Cleaning remote {0} {1} build.",
     "_CleaningRemoteResources.comment": "Message printed when building --clean for a remote target. {0} is the platform, and {1} is release/debug",
 
-    "CordovaCommandFailed": "'cordova {1}' failed with exit code {0}",
-    "_CordovaCommandFailed.comment": "Error message passed back after invoking Cordova fails. {0} is the exit code, and {1} is the arguments passed to it",
-
-    "CordovaCommandFailedWithError": "'cordova {0}' failed.",
-    "_CordovaCommandFailedWithError.comment": "Error message passed back after invoking Cordova fails. {0} is the exit code, and {1} is the arguments passed to it",
-
-    "CordovaCommandUnhandledException": "Cordova threw an unhandled exception, check http://go.microsoft.com/fwlink/?LinkId=623759 for possible details.",
-    "_CordovaCommandUnhandledException.comment": "Error message reported when Cordova throws an unhandled exception",
-
-    "CordovaCmdNotFound": "Cordova not found, please run 'npm install –g cordova' to install Cordova globally",
-    "_CordovaCmdNotFound.comment": "Error message shown when no Cordova is found on the machine",
     "UnimplementedAbstractMethod": "This is an abstract method and should be implemented by a derived overload",
     "_UnimplementedAbstractMethod.comment": "Error thrown for an unimplemented method",
