+++ src/taco-tests-utils/tacoErrorTestHelper.ts
@@ -15,13 +15,9 @@ import shouldModule = require("should");
 
 export class TacoErrorTestHelper {
     public static verifyTacoErrors(fileName: string, resources: any, minErrorCode: number, maxErrorCode: number): void {
-        var errorCodes: any = require(fileName);
-        // for taco-utils, we have TacoErrorCode is nested under module
-        if (errorCodes.TacoErrorCode) {
-            errorCodes = errorCodes.TacoErrorCode;
-        }
+        var errorCodes: any = TacoErrorTestHelper.getErrorCodes(fileName);
 
-        Object.keys(errorCodes).forEach(function (errorCode: string): void {
+        Object.keys(errorCodes).forEach(function(errorCode: string): void {
             // Loop over all error codes, filter out numeric values
             if (isNaN(parseInt(errorCode, 10))) {
                 // Verify that we don't have 2 errors with same error code
@@ -37,5 +33,21 @@ export class TacoErrorTestHelper {
                 shouldModule(resources.getString(errorCode)).not.equal(null, "no resources found for error code " + errorCode);
             }
         });
+    }
+
+    public static verifyExcludedTacoErrors(fileName: string, resources: any, excludedErrorCode: number): void {
+        var errorCodes: any = TacoErrorTestHelper.getErrorCodes(fileName);
+        var excludedErrorCodeStr: string = errorCodes[excludedErrorCode];
+        shouldModule(excludedErrorCodeStr).equal(undefined, "Exclude error code " + excludedErrorCode + "shouldn't be present");
+    }
+
+    private static getErrorCodes(fileName: string): any {
+        var errorCodes: any = require(fileName);
+        // for taco-utils, we have TacoErrorCode is nested under module
+        if (errorCodes.TacoErrorCode) {
+            errorCodes = errorCodes.TacoErrorCode;
+        }
+
+        return errorCodes;
     }
 }
