+++ src/taco-tests-utils/tacoTestsUtils.ts
@@ -8,10 +8,11 @@
 
 "use strict";
 import nodeFakes = require("./nodeFakes");
-import projectHelper = require("./projectHelper");
+import projectHelper = require("./testProjectHelper");
 import telemetryFakes = require("./telemetryFakes");
 import memoryStream = require("./memoryStream");
 import tacoErrorTestHelper = require("./tacoErrorTestHelper");
+import mockCordova = require("./mockCordova");
 
 module TacoTestsUtils {
     /* tslint:disable:variable-name */
@@ -22,6 +23,7 @@ module TacoTestsUtils {
     export var ProjectHelper: typeof projectHelper.ProjectHelper = projectHelper.ProjectHelper;
     export var MemoryStream: typeof memoryStream.MemoryStream = memoryStream.MemoryStream;
     export var TacoErrorTestHelper: typeof tacoErrorTestHelper.TacoErrorTestHelper = tacoErrorTestHelper.TacoErrorTestHelper;
+    export var MockCordova: typeof mockCordova.MockCordova = mockCordova.MockCordova;
     /* tslint:enable:variable-name */
 }
