+++ src/taco-utils/package.json
@@ -33,6 +33,7 @@
         "semver": "2.3.1",
         "unorm": "1.3.3",
         "windows-no-runnable": "0.0.6",
+        "wrench": "^1.5.8",
         "lodash": "^3.10.1"
     },
     "devDependencies": {
