+++ src/taco-utils/tacoUtils.ts
@@ -11,6 +11,8 @@ import argsHelper = require ("./argsHelper");
 import buildInfo = require ("./buildInfo");
 import commands = require ("./commands");
 import cordovaConfig = require ("./cordovaConfig");
+import cordovaHelper = require ("./cordovaHelper");
+import cordovaWrapper = require ("./cordovaWrapper");
 import countStream = require ("./countStream");
 import resourceManager = require ("./resourceManager");
 import helpCommandBase = require ("./helpCommandBase");
@@ -20,6 +22,7 @@ import logger = require ("./logger");
 import loggerHelper = require ("./loggerHelper");
 import logLevel = require ("./logLevel");
 import newlineNormalizerStream = require ("./newlineNormalizerStream");
+import projectHelper = require ("./projectHelper");
 import processLogger = require ("./processLogger");
 import promisesUtils = require ("./promisesUtils");
 import tacoError = require ("./tacoError");
@@ -40,6 +43,8 @@ module TacoUtility {
     export var BuildInfo: typeof buildInfo.BuildInfo = buildInfo.BuildInfo;
     export var Commands: typeof commands.Commands = commands.Commands;
     export var CordovaConfig: typeof cordovaConfig.CordovaConfig = cordovaConfig.CordovaConfig;
+    export var CordovaHelper: typeof cordovaHelper.CordovaHelper = cordovaHelper.CordovaHelper;
+    export var CordovaWrapper: typeof cordovaWrapper.CordovaWrapper = cordovaWrapper.CordovaWrapper;
     export var CountStream: typeof countStream.CountStream = countStream.CountStream;
     export var Logger: typeof logger.Logger = logger.Logger;
     export var LoggerHelper: typeof loggerHelper.LoggerHelper = loggerHelper.LoggerHelper;
@@ -49,6 +54,7 @@ module TacoUtility {
     export var InstallLogLevel: typeof installLogLevel.InstallLogLevel = installLogLevel.InstallLogLevel;
     export var JsonSerializer: typeof jsonSerializer.JsonSerializer = jsonSerializer.JsonSerializer;
     export var ProcessLogger: typeof processLogger.ProcessLogger = processLogger.ProcessLogger;
+    export var ProjectHelper: typeof projectHelper.ProjectHelper = projectHelper.ProjectHelper;
     export var ResourceManager: typeof resourceManager.ResourceManager = resourceManager.ResourceManager;
     export var TacoError: typeof tacoError.TacoError = tacoError.TacoError;
     export var TacoErrorLevel: typeof tacoError.TacoErrorLevel = tacoError.TacoErrorLevel;
