+++ src/typings/tacoJsonEditParams.d.ts
@@ -8,10 +8,8 @@
 
 /// <reference path="../typings/node.d.ts" />
 
-declare module TacoTestsUtils {
-    class TacoErrorTestHelper {
-        public static verifyTacoErrors(fileName: string, resources: any, minErrorCode: number, maxErrorCode: number): void;
-        public static verifyExcludedTacoErrors(fileName: string, resources: any, excludedErrorCode: number): void;
-    }
+interface ITacoJsonEditParams {
+    projectPath: string;
+    isKitProject: boolean;
+    version: string;
 }
-
