+++ src/typings/tacoProjectInfo.d.ts
@@ -8,10 +8,9 @@
 
 /// <reference path="../typings/node.d.ts" />
 
-declare module TacoTestsUtils {
-    class TacoErrorTestHelper {
-        public static verifyTacoErrors(fileName: string, resources: any, minErrorCode: number, maxErrorCode: number): void;
-        public static verifyExcludedTacoErrors(fileName: string, resources: any, excludedErrorCode: number): void;
-    }
+interface IProjectInfo {
+    isTacoProject: boolean;
+    cordovaCliVersion: string;
+    configXmlPath: string;
+    tacoKitId?: string;
 }
-
