+++ src/typings/tacoUtils.d.ts
@@ -10,6 +10,8 @@
 /// <reference path="../typings/buildInfo.d.ts" />
 /// <reference path="../typings/commands.d.ts" />
 /// <reference path="../typings/cordovaConfig.d.ts" />
+/// <reference path="../typings/cordovaHelper.d.ts" />
+/// <reference path="../typings/cordovaWrapper.d.ts" />
 /// <reference path="../typings/countStream.d.ts" />
 /// <reference path="../typings/jsonSerializer.d.ts" />
 /// <reference path="../typings/logger.d.ts" />
@@ -17,6 +19,7 @@
 /// <reference path="../typings/logLevel.d.ts" />
 /// <reference path="../typings/newlineNormalizerStream.d.ts" />
 /// <reference path="../typings/processLogger.d.ts" />
+/// <reference path="../typings/projectHelper.d.ts" />
 /// <reference path="../typings/resourceManager.d.ts" />
 /// <reference path="../typings/tacoError.d.ts" />
 /// <reference path="../typings/tacoGLobalConfig.d.ts" />
